requirements.txt
packaging<24
urllib3<2.1
limits<4.0

psycopg2-binary>=2.9.9

chromadb==0.4.24

selenium>=4.18.0
undetected-chromedriver>=3.5.4

groq>=0.4.1

python-dotenv>=1.0.0